+++ src/App.tsx
@@ -2,15 +2,10 @@ import MainPage from "pages/MainPage";
 import CardRegisterForm from "pages/RegisterPage/CardRegisterForm";
 import LastPage from "pages/LastPage";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
-import { useState } from "react";
-import { INITIAL_CARD_INFO } from "constants/initialCardInfo";
-import { createContext } from "react";
-import { CardInfoState } from "types";
 import GotLost from "pages/GotLost";
+import CardInfoProvider from "components/CardInfoProvider";
 
 const App = () => {
-  const [cardInfo, setCardInfo] = useState(INITIAL_CARD_INFO);
-
   return (
     <BrowserRouter basename={process.env.PUBLIC_URL}>
       <Routes>
@@ -18,16 +13,19 @@ const App = () => {
         <Route
           path="/register"
           element={
-            <>
-              <CardInfoContext.Provider
-                value={{ cardInfo: cardInfo, setCardInfo: setCardInfo }}
-              >
-                <CardRegisterForm allCardInfo={cardInfo} />
-              </CardInfoContext.Provider>
-            </>
+            <CardInfoProvider>
+              <CardRegisterForm />
+            </CardInfoProvider>
           }
         />
-        <Route path="/completion" element={<LastPage cardInfo={cardInfo} />} />
+        <Route
+          path="/completion"
+          element={
+            <CardInfoProvider>
+              <LastPage />
+            </CardInfoProvider>
+          }
+        />
         <Route path="/got-lost" element={<GotLost />} />
         <Route path="*" element={<Navigate replace to="/got-lost" />} />
       </Routes>
@@ -36,8 +34,3 @@ const App = () => {
 };
 
 export default App;
-
-export const CardInfoContext = createContext<CardInfoState>({
-  cardInfo: INITIAL_CARD_INFO,
-  setCardInfo: () => {},
-});
