+++ src/pages/RegisterPage/FormInputs/SecurityCodeInput.tsx
@@ -3,8 +3,8 @@ import styled from "styled-components";
 import { changeInvalidValueToBlank } from "utils/inputValidator";
 import Input, { SecurityCodeInputStyle } from "components/Input";
 import { CodeCaption } from "components/CaptionStyle";
+import { CardInfoContext } from "components/CardInfoProvider";
 import { LIMIT_LENGTH, VALID_INPUT } from "constants/limit";
-import { CardInfoContext } from "App";
 const { ONLY_NUMBER } = VALID_INPUT;
 
 const SecurityCodeInput = () => {
